Renamed MenuGroup.hooks to MenuHeading.hooks
