+++ packages/components/src/Menu/MenuGroupLabel.tsx
@@ -27,7 +27,7 @@
 import React, { FC, useRef, RefObject } from 'react'
 import styled from 'styled-components'
 import { Heading } from '../Text/Heading'
-import { useElementVisibility } from './MenuGroup.hooks'
+import { useElementVisibility } from './MenuHeading.hooks'
 
 export const MenuGroupLabel: FC = ({ children }) => {
   const labelShimRef: RefObject<any> = useRef()
